Move GrafikElementFormAdapter to feature layer
